ref: Change typing to d.ts - 06
