+++ src/types/shared/errors.d.ts
