+++ src/types/shared/file.d.ts
