--- src/types/users/GameInfo.ts
@@ -1,1 +1,0 @@
-export type GameInfoOptions = 'badges' | 'campaigns' | 'characters';
