--- src/types/users/Mock.ts
@@ -1,9 +1,0 @@
-export default interface Mock {
-    instance: {
-        _id?: string;
-        active?: boolean;
-        en?: unknown;
-        pt?: unknown;
-    };
-    description: string;
-}
